fixed send event message use case
